Change flipped background handling
Previously the background of all panoramas was flipped by setting the x-scale on the panorama to -1, in order to correctly flip the texture when viewed from inside the geometry. This also affected all the _children_ of the panorama, which was not desirable and required some strange handling code where all children (except Infospots) were wrapped with _another_ x-inverted wrapper to undo the scaling.

This change undoes the inverted x at the panorama level by instead making the background geometry a separate object and inverting the x-scale on that. The panorama itself is now just an empty Object3D with normal scaling and all children are treated normally.